+++ Source/Algorithm/ChaCha20/ChaCha20.hpp
@@ -24,8 +24,8 @@ public:
 private:
 	// The ChaCha20 algorithm is a stream cipher using a key of 256 bit size.
 	// the ChaCha20 stream cipher does not have a block operation mode therefore it does not use padding.
-	std::unique_ptr<CryptoPP::ChaCha::Encryption> m_chacha_encryption; // ChaCha20 encryptor
-	std::unique_ptr<CryptoPP::ChaCha::Decryption> m_chacha_decryption; // ChaCha20 decryptor
+	std::unique_ptr<CryptoPP::ChaCha::Encryption> m_chacha_encryptor; // ChaCha20 encryptor
+	std::unique_ptr<CryptoPP::ChaCha::Decryption> m_chacha_decryptor; // ChaCha20 decryptor
 };
 
 NS_ENIGMA_END