+++ Source/Config/Config.cpp
@@ -4,8 +4,11 @@
 NS_ENIGMA_BEGIN
 
 Config::Config(const String& file_name) noexcept(false)
+	:
+	m_ini_reader(new INIReader(file_name))
 {
-	m_ini_reader = std::make_unique<INIReader>(file_name);
+	ENIGMA_ASSERT_OR_THROW(m_ini_reader, "Failed to initialize INIReader");
+
 	if (m_ini_reader->ParseError() < 0) 
 	{
 		throw std::runtime_error("Couldn't load config file " + file_name);