+++ Source/Core/Macros.hpp
@@ -27,7 +27,9 @@
 //
 
 
-
+/// Unused macro helper
+#define ENIGMA_UNUSED(v) ((void)(v))
+///
 
 ///Asserts
 #ifdef ENIGMA_ENABLE_ASSERTS