+++ Source/Utility/Random.hpp
@@ -10,7 +10,10 @@ class ENIGMA_API Random
 {
 public:
 	/*
-	*	Returns a Random Real between min and max
+	*	@brief Generates a random real between a range
+	*	@param min: minimum value of range (included)
+	*	@param max: maximum value of range (included)
+	*	@return Random real between min and max
 	*/
 	template<typename T>
 	static typename std::enable_if<std::is_floating_point<T>::value, T>::type
@@ -22,7 +25,10 @@ public:
 	}
 
 	/*
-	*	Returns a Random Integer between min and max
+	*	@brief Generates a random integer between a range
+	*	@param min: minimum value of range (included)
+	*	@param max: maximum value of range (included)
+	*	@return Random integer between min and max
 	*/
 	template<typename T>
 	static typename std::enable_if<std::is_integral<T>::value && !std::is_same<T, bool>::value, T>::type
@@ -34,8 +40,9 @@ public:
 	}
 
 	/*
-	*	Returns a Random bool value, either 'true' or 'false'
+	*	@brief Generates a random boolean value
 	*	@param p: Probability of true (50% by default)
+	*	@return Returns a Random bool value, either 'true' or 'false'
 	*/
 	static bool Bool(const f64 p = 0.5) noexcept
 	{
@@ -44,14 +51,16 @@ public:
 	}
 
 	/*
-	*	Returns a Random String(length) consists of alphanumeric characters
+	*	@brief Generates a random string with length
+	*	@param length: length of the generated random string
+	*	@return Returns a Random String with specified length consists of alphanumeric characters and special characters
 	*/
 	static String Str(const size_t length) noexcept
 	{
 		String str(length, '\000');
 		for (char& c : str)
 		{
-			const i16 i = Int(static_cast<i16>(1), static_cast<i16>(3));
+			const i16 i = Int(static_cast<i16>(1), static_cast<i16>(4));
 			switch (i)
 			{
 			case 1: // a-z
@@ -62,6 +71,10 @@ public:
 				break;
 			case 3: // 0-9
 				c = Int(static_cast<i16>('0'), static_cast<i16>('9')); // ascii 0 = 48 | 9 = 57
+				break;
+			case 4: // special characters
+				using Constants::Algorithm::SPECIAL_CHARACTERS;
+				c = SPECIAL_CHARACTERS[Int(static_cast<size_t>(0), ENIGMA_ARRAY_SIZE(SPECIAL_CHARACTERS))];
 				break;
 			}
 		}